add JDBC Database connection settings to application.properties file

+++ src/main/java/net/ed/api/controller/JSONReader.java
@@ -26,8 +26,8 @@ public class JSONReader {
 		BigDecimal high = null;
 		BigDecimal low = null;
 		BigDecimal close = null;
-		BigDecimal volume = null;
-		Double unadjustedVolume = null;
+		int volume = (Integer) null;
+		int unadjustedVolume = (Integer) null;
 		BigDecimal change = null;
 		Double changePercent = null;
 		BigDecimal vwap = null;
